+++ swarm/CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.42.0 [unreleased]
 
+- Remove uncontructed variant `Timer` from `ConnectionHandlerUpgrErr`.
+
 - Update to `libp2p-core` `v0.39.0`.
 
 - Removed deprecated Swarm constructors. For transition notes see [0.41.0](#0.41.0). See [PR 3170].